Using double for the first argument avoid undefined behaviour. Closes #296

+++ core/Scope.cpp
@@ -139,7 +139,8 @@ void Scope::log(const float* values){
 	postlog();
 
 }
-void Scope::log(float chn1, ...){
+
+void Scope::log(double chn1, ...){
 	
 	if (!prelog()) return;
